tests: reorganize test utils for plank
